requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==4.2.3
django-crispy-forms==2.0
django-mathfilters==1.0.0
django-rosetta==0.9.9
django-translated-fields==0.12.0
django-utils-six==2.0
idna==3.4
Pillow==10.0.0
polib==1.2.0
python-gettext==5.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
